Move `internal.labels` to `format` package for use by providers
