+++ controlplane/kubeadm/internal/cluster_labels.go
@@ -19,7 +19,7 @@ package internal
 import (
 	clusterv1 "sigs.k8s.io/cluster-api/api/v1beta1"
 	controlplanev1 "sigs.k8s.io/cluster-api/controlplane/kubeadm/api/v1beta1"
-	capilabels "sigs.k8s.io/cluster-api/internal/labels"
+	"sigs.k8s.io/cluster-api/util/labels/format"
 )
 
 // ControlPlaneMachineLabelsForCluster returns a set of labels to add to a control plane machine for this specific cluster.
@@ -36,6 +36,6 @@ func ControlPlaneMachineLabelsForCluster(kcp *controlplanev1.KubeadmControlPlane
 	labels[clusterv1.ClusterNameLabel] = clusterName
 	labels[clusterv1.MachineControlPlaneLabel] = ""
 	// Note: MustFormatValue is used here as the label value can be a hash if the control plane name is longer than 63 characters.
-	labels[clusterv1.MachineControlPlaneNameLabel] = capilabels.MustFormatValue(kcp.Name)
+	labels[clusterv1.MachineControlPlaneNameLabel] = format.MustFormatValue(kcp.Name)
 	return labels
 }
